Switch to using ls instead of git in test.
When running in the docker container used by GitHub Actions, using git
causes the test to fail because git returns a non-zero error code. This
is because the user that the git command is run under is different than
the user that owns the repo files in the container. Using ls avoids this
issue while still staying true to the test's purpose.