Add time limit for safety reasons.
PyShell's test suite currently does not take very long to execute (~2
minutes). This means that GitHub's default timeout of 6 hours is
unnecessarily long and would burn through far more CI/CD minutes than
necessary. The timeout set in this commit is a bit short, but it's being
set with the expectation that it'll be raised over time as PyShell's
test suite expands.